requirements.txt
playwright
pluggy
py
pyee
pyparsing
pytest
pytest-base-url
pytest-forked
pytest-playwright
pytest-xdist
python-slugify
requests
text-unidecode
toml
tomli
urllib3
websockets
python-dotenv
jsonify
flask
